+++ packages/core/src/connectivity/scp-cf/destination/destination-service.ts
@@ -6,7 +6,7 @@ import {
 import CircuitBreaker from 'opossum';
 import axios, { AxiosRequestConfig, AxiosResponse, AxiosError } from 'axios';
 import { decodeJwt, wrapJwtInHeader } from '../jwt';
-import { getAgentConfig } from '../../../http-client';
+import {getAgentConfig, urlAndAgent} from '../../../http-client';
 import {
   circuitBreakerDefaultOptions,
   ResilienceOptions
@@ -15,8 +15,7 @@ import { CachingOptions } from '../cache';
 import {
   addProxyConfigurationInternet,
   ProxyStrategy,
-  proxyStrategy,
-  urlAndAgent
+  proxyStrategy
 } from '../proxy-util';
 import { parseDestination } from './destination';
 import { Destination, DestinationType } from './destination-service-types';
