+++ packages/core/src/http-agent/http-agent.ts
@@ -1,14 +1,18 @@
 import * as http from 'http';
 import * as https from 'https';
-import { assoc, createLogger, last } from '@sap-cloud-sdk/util';
+import {assoc, createLogger, last} from '@sap-cloud-sdk/util';
 import {
+  addProxyConfigurationInternet,
   Destination,
   DestinationCertificate,
+  getProtocolOrDefault,
   Protocol,
   proxyAgent,
-  getProtocolOrDefault
+  ProxyStrategy,
+  proxyStrategy
 } from '../connectivity/scp-cf';
-import { HttpAgentConfig, HttpsAgentConfig } from './agent-config';
+import {HttpAgentConfig, HttpsAgentConfig} from './agent-config';
+import {HttpRequestConfig} from "../http-client";
 
 const logger = createLogger({
   package: 'core',
@@ -171,4 +175,24 @@ export function getUrlProtocol(destination: Destination): Protocol | undefined {
       return urlParts[0] as Protocol;
     }
   }
+}
+
+/**
+ * Builds part of the request config containing the URL and if needed proxy agents or normal http agents.
+ * Considers the NO_Proxy env variable together with the targetUri.
+ * @param targetUri used as baseURL in request config
+ * @returns HttpRequestConfig containing baseUrl and http(s) agents.
+ */
+export function urlAndAgent(
+    targetUri: string
+): Pick<HttpRequestConfig, 'baseURL' | 'httpAgent' | 'httpsAgent' | 'proxy'> {
+  let destination: Destination = {url: targetUri, proxyType: 'Internet'};
+  if (proxyStrategy(destination) === ProxyStrategy.INTERNET_PROXY) {
+    destination = addProxyConfigurationInternet(destination);
+  }
+  return {
+    baseURL: destination.url,
+    proxy: false,
+    ...getAgentConfig(destination)
+  };
 }
