+++ packages/core/src/http-client/index.ts
@@ -1,6 +1,6 @@
 export * from './http-client';
 export * from './http-client-types';
-export * from './http-agent';
-export * from './agent-config';
+export * from '../http-agent/http-agent';
+export * from '../http-agent/agent-config';
 export * from './http-request-config';
 export { buildCsrfHeaders } from './csrf-token-header';
